just dfoing a bit of cleanup related to the bifurcation refactor and moving some types around so things are cleaner and easier to read
